backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
pillow==10.1.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
